requirements.txt
google-cloud-bigquery==1.10.0
python-dateutil==2.8.0
numpy==1.16.3
pandas==0.24.2
pyod==0.6.8
pyarrow==0.13.0
scikit-learn==0.20.3
xgboost==0.82
Click==7.0
